Default templates initializer - skip on db:migrate

+++ config/initializers/z_12_default_templates.rb
@@ -4,20 +4,27 @@
 # least can see an entry / how they look / how they work.
 
 
-# --------------------------------------------------------------- Note template
-if !NoteTemplate.pwd.exist?
-  NoteTemplate.pwd.mkpath
-  NoteTemplate.new(name: 'Basic fields', content: "#[Title]#\n\n\n#[Description]#\n\n").save
-end
+# Make sure this is not the first time we're init'ing the app via rake
+# db:migrate. If we don't have a `configurations` table, we can't know
+# where the templates are stored!
+if Configuration.table_exists?
 
-
-
-# ----------------------------------------------------------------- Methodology
-if !Methodology.pwd.exist?
-  Methodology.pwd.mkpath
-  %w{sample.xml}.each do |file|
-    source_file      = Rails.root.join('spec', 'fixtures', 'files', 'methodologies', file)
-    destination_file = Methodology.pwd.join(file)
-    FileUtils.cp(source_file, destination_file)
+  # --------------------------------------------------------------- Note template
+  if !NoteTemplate.pwd.exist?
+    NoteTemplate.pwd.mkpath
+    NoteTemplate.new(name: 'Basic fields', content: "#[Title]#\n\n\n#[Description]#\n\n").save
   end
+
+
+
+  # ----------------------------------------------------------------- Methodology
+  if !Methodology.pwd.exist?
+    Methodology.pwd.mkpath
+    %w{sample.xml}.each do |file|
+      source_file      = Rails.root.join('spec', 'fixtures', 'files', 'methodologies', file)
+      destination_file = Methodology.pwd.join(file)
+      FileUtils.cp(source_file, destination_file)
+    end
+  end
+
 end
